Enable Vim to see CTRL-S keystrokes.
The Vim plugin I'm writing remaps Colemak to QWERTY in Normal, Visual,
and Operator-pending modes so that commands and motions use the same
physical keystrokes. One of the commands it will remap is CTRL-S, which
corresponds to CTRL-D.

However, the OS X terminal driver has software flow control enabled by
default, and it intercepts CTRL-S as a control signal. Disabling this
allows Vim to see CTRL-S.