Have Vim NOT set Terminal.app's proxy icon.
This reverts commit 9c0e52912dcc628eee57c5844adf47b3eddef556 "Have Vim
set Terminal.app's proxy icon".

Caught up in the novelty of setting the proxy icon, I forgot that the
icon is meant to represent the current working directory. Setting it to
a file breaks opening new tabs or windows to the CWD (having "New
windows open with" or "New tabs open with" set to "Same Working
Directory" in Terminal Preferences).