Create master setup script and utility library.
The first step in creating a kind of "installer" for my Unix
configuration. The master script will find and source any other "setup"
files located in the local directory tree.

The utility library is meant to implement common functionality in
a POSIX-compliant manner. So far, it provides a function that accepts
relative paths and creates absolute-path symlinks.