setup: Introduce new, simpler linking function.
The old linking function, CreateAbsoluteSymlinks, was getting kind of
unwieldy and hard-to-understand, from a functional standpoint. At least
as far as I was concerned.

The new one, ln_abs, should be less of a problem. It behaves
more-or-less just like ln(1), except that it only does "one-to-one"
linking (the "first synopsis form" from the POSIX man page[*]) and, of
course, resolves absolute canonical pathnames before linking. It accepts
both the -f and -s options.

    [*] http://pubs.opengroup.org/onlinepubs/009695399/utilities/ln.html