setup: Fix option-processing bug in ln_abs.
As getopts processes options, it maintains the index of the next
argument to be processed in the shell variable OPTIND. This variable is
incremented every time a getopts invocation processes an option
argument. However, getopts only increments OPTIND; it never resets it.
(Shells set OPTIND to 1 on invocation.) And OPTIND, as a shell variable,
is shared between all function calls within a single shell session.

This led to a bug wherein the first call to ln_abs during a shell
invocation would see its options just fine, but any subsequent calls
would get no options at all. We fix this by resetting OPTIND on every
call to ln_abs.