Convert all leading underscores in filenames
One convention I use in this repository is to use leading underscores in
filenames and convert them to periods on installation. This prevents
*nix tools from treating them as hidden.

This commit fixes a bug wherein this conversion only happened in the
path component immediately following the module name. That is,
a repository file "foo/_bar/_baz" would install to "$(prefix)/.bar/_baz"
instead of "$(prefix)/.bar/.baz" as intended.