macports: Fix path augmentation
Thanks to my inadequate testing, I had assumed that the glob qualifiers
applied to the whole array. Turns out, they only apply to the last
member because filename generation happens after array expansion, and
the qualifiers get glued to the last member. Hence, that was the only
one that would be filtered out if invalid:

    % foo=(/foo /usr /usr/local /bar)
    % echo $foo(N)
    /foo /usr /usr/local

Using ^-expansion distributes the qualifiers to every array member:

    % echo ${^foo}(N)
    /usr /usr/local