zsh: Verify autoloaded functions properly
I had mistakenly assumed that autoload would fail if it couldn't find
the requested function, but it actually successfully defines a skeleton
function that itself calls "autoload -X", which *does* fail:

    % autoload nonexistent-func; echo $?
    0
    % functions nonexistent-func
    nonexistent-func () {
    	# undefined
    	builtin autoload -X
    }
    % nonexistent-func; echo $?
    zsh: nonexistent-func: function definition file not found
    1

So a better way to verify that an autoloaded function exists is to try
to use it. Surprise!

(Aside: I should really start testing things instead of just guessing.)