lynx: Defer inclusion of the local configuration
This file previously included a global configuration at the beginning,
allowing subsequent settings to override global ones. When I stopped
using the global configuration and started allowing for a per-machine
one, I reused the INCLUDE directive in place, but it belongs at the end
so that the per-machine configuration takes ultimate precedence.