zsh: Set colors with %F prompt escape (again!)
I used to set colors with the %F prompt escape but switched to ANSI
escape sequences for compatibility with old versions of zsh [1].
Unfortunately, raw ANSI sequences are not portable across all terminals
[2][3]. The recommended method -- substituting tput(1) output -- would
be hard to apply robustly to completion formats, which aren't evaluated
the same way prompt strings are [4]. Instead of continuing to dig this
hole, use %F if it's supported and color-free alternatives otherwise.

One wrinkle is that the "named" form of %F (e.g., '%F{red}') is actually
implemented using ANSI sequences [5]! Stick with the "numbered" form
(e.g., '%1F'), which uses termcap.

  [1]: https://github.com/larryv/dotfiles/commit/f62256ed0f058d3005bb029a990ced21f92c55f3
  [2]: http://wiki.bash-hackers.org/scripting/terminalcodes
  [3]: http://mywiki.wooledge.org/BashFAQ/037
  [4]: http://mywiki.wooledge.org/BashFAQ/053
  [5]: https://sourceforge.net/p/zsh/code/ci/zsh-4.3.7/tree/Src/prompt.c