zsh: Send full hostname to Terminal
I've been truncating the hostname ever since I started invoking
hostname(1), but I have no idea why. If that ever worked, it has not
done so in some time, causing Terminal's proxy icon to be overlaid with
a question mark and disabling the icon's menu. This tweak fixes the
situation on macOS 10.14.1 Mojave.