zsh: Send login name to Terminal
The Timing application recommends including the login name in Terminal
titles to maximize its chances of correctly extracting the current
working directory (https://timingapp.com/help/terminal). It already
seems to work fine without this change, but the change doesn't cause
Terminal any problems, so why not.

Slashes must be percent-encoded in the userinfo and host but should not
be encoded in the path, so handle each part separately.