zsh: Make hidden files in .zsh visible
It doesn't make much sense for the zsh startup files to be hidden while
being located in the .zsh directory, which is already hidden. Most other
files in that directory are not hidden, including the topic startup
files. (This thought did occur to me when I created .zsh, but for some
reason all I did was rename the history file [*].)

Rename the startup files without leading dots and create symbolic links
at the old names so zsh can still find them. Use a non-dotted name for
the completion system's dumped configuration, instead of the default
".zcompdump".

    [*]: commit d739a681189b04e83f7effc3e466968177f4b4e8