zsh: Mimic the way .profile sources topic scripts
When looking for topic scripts to source, only consider regular files
and symbolic links that point to regular files. This is for consistency
with .profile, which looks for topic scripts this way:

    for script in ~/.profile.d/*.sh; do
        if [ -f "$script" ]; then
            . "$script"
        fi
    done

The conditional is necessary to protect against the glob matching
nothing and expanding to itself, but it doubles as a sensible filter.
I can't imagine why a topic script would ever be anything other than
a regular file or a symlink to one.