zsh: Ditch anonymous functions (again...)
This is neither a renouncement of b608604 (zsh: Make startup files'
temporary variables local, 2020-06-06) nor an admission that 29eb1e5
(java: Ditch the anonymous zsh function, 2016-08-03) was right. It is
something completely different!

Anonymous functions were implemented in zsh 4.3.7 [1]. Use alternatives
to improve backward compatibility with Mac OS X 10.5 Leopard, which
ships with zsh 4.3.4.

When using regular functions for local sh emulation, work around a bug
fixed in 4.3.7 [2] wherein the emulation mode is not restored upon
exiting a function in which `emulate -L` had changed it.

  [1]: https://www.zsh.org/mla/workers/2008/msg00915.html (workers/25255)
  [2]: https://www.zsh.org/mla/users/2008/msg00785.html (users/13169)