macports: Rewrite and generalize PATH modification
I know, 3e4d3c3 (macports: Overhaul PATH modification, 2020-06-05) was
like two months ago. I would amend it had I not pushed it already.

Implement a new algorithm for promoting elements in a colon-delimited
list. It improves on the old one by handling duplicates and being easier
to reason about and extend to cover edge cases. (On the other hand, it
might be a little slower, but that's not important here.)

Break it out into a general function that can be used to adjust lists
other than PATH. Apply it to /etc/paths.d/macports using sed(1) and an
ad hoc xargs(1)-like function that can invoke other functions.