build: Replace GNU make functions portably
This is step 6 of 9 of the POSIX make rewrite.

Functions are not standardized in POSIX.1-2017 [*], so replace them all
portably. Unfortunately, this means defining all the rules manually.
Make the most of this by implementing optimizations that, under the old
metaprogramming regime, were more trouble than they were worth (e.g.,
installing multiple files with one command, setting ad hoc permissions
with `$(INSTALL) -m` instead of chmod(1), etc.).

Although not strictly necessary, accrue the prerequisites of the "all"
target in the submakefiles instead of the main one, to match the other
convenience targets. Remove the now-purposeless MODULES variable.

  [*]: https://pubs.opengroup.org/onlinepubs/9699919799/utilities/make.html