build: Try to replace .DELETE_ON_ERROR portably
This is step 8 of 9 of the POSIX make rewrite.

Remove .DELETE_ON_ERROR because it is not standardized as a special
target in POSIX.1-2017 [1].

According to the manual for GNU make, dated 2020-01-19 [2]:

    Usually when a recipe line fails, if it has changed the target
    file at all, the file is corrupted and cannot be used -- or at
    least it is not completely updated. Yet the file's time stamp
    says that it is now up to date, so the next time `make` runs, it
    will not try to update that file. [...] So generally the right
    thing to do is to delete the target file if the recipe fails
    after beginning to change the file. `make` will do this if
    `.DELETE_ON_ERROR` appears as a target. This is almost always
    what you want `make` to do, but it is not historical practice;
    so for compatibility, you must explicitly request it.

Approximate this behavior portably by explicitly deleting targets if any
commands that update them exit with a nonzero status. Do this regardless
of whether their timestamps have changed because I can't be bothered.

(Currently this only applies to the rule that processes M4 templates.)

  [1]: https://pubs.opengroup.org/onlinepubs/9699919799/utilities/make.html#tag_20_76_13_04
  [2]: https://www.gnu.org/software/make/manual/html_node/Errors