build: Rename makefile to "Makefile" and add .POSIX
This is step 9 of 9 of the POSIX make rewrite.

The previous steps strove to make the build system highly portable by
removing all features not explicitly standardized in POSIX.1-2017. Now
that the main makefile has a standard name [1] and begins [2] with the
.POSIX special target [3], it should produce strict POSIX behavior in
compliant make implementations. (*knocks on wood*)

Good work, team!

  [1]: https://pubs.opengroup.org/onlinepubs/9699919799/utilities/make.html#tag_20_76_13_01
  [2]: https://pubs.opengroup.org/onlinepubs/9699919799/utilities/make.html#tag_20_76_13_04
  [3]: https://www.gnu.org/software/make/manual/html_node/Special-Targets#index-_002ePOSIX