sh, zsh: Create the topic script directories
Previously, every module that contained a shell component independently
ensured that the necessary topic directories existed; this was a weird
artifact of the previous build system design. Now, the topic directories
are the responsibility of the sh and zsh modules, which ensure that they
exist with their respective *-installdirs targets. Other modules should
depend on those targets as needed. Consolidating management of the topic
directories clarifies modules' responsibilities and relationships.