tmux: Fix misquoted "default-command" definition
This has been broken since b25eb49 (tmux: Apply OS X pasteboard fix more
cleanly, 2016-05-13). I must have thought that the double and single
quotes nested and would produce this value for "default-command", to be
evaluated by a shell:

    exec reattach-to-user-namespace -l "$SHELL"

In fact, the double quotes around $SHELL resulted in two disjoint quoted
strings and left $SHELL to be expanded and unsafely injected by tmux:

    exec reattach-to-user-namespace -l /opt/local/bin/zsh

Fix it by replacing the outer pair of double quotes with braces, which
were added to tmux 3.0 [*] for exactly this sort of thing.

  [*]: https://raw.githubusercontent.com/tmux/tmux/3.0/CHANGES