tmux: Try to set "default-terminal" intelligently
Operative word is "try".

Test whether the declared terminal has 256-color support instead of
blindly using a *-256color terminal description. Try to make sure the
chosen description is available before locking it in.

Brace-quoting makes it tractable to deeply nest "if-shell" commands but
limits me to tmux 3.0 [*], which I deem worth it.

  [*]: https://raw.githubusercontent.com/tmux/tmux/3.0/CHANGES