zsh: Properly pass escape sequences through tmux
I just found out that there is official documentation [1] on how to
properly package escape sequences so tmux passes them through to the
terminal. It is much better than that random mailing list post [2].

Package the original sequence correctly by doubling its $'\e' characters
and prefixing it with $'\ePtmux;' instead of $'\ePtmux;\e'. (The extra
$'\e' was actually a double of the original sequence's leading $'\e'.)

Define the sequences using $'...' quoting instead of backslash escapes
inside '...'. This simplifies $'\e'-doubling because the variables now
contain actual $'\e' characters, which are trivial to search for (as
opposed to '\e' preceded by an even number of '\' characters).

Include the URI in the sequence from the beginning to clarify that it is
in principle not exempt from the $'\e'-doubling requirement. (Since it
is percent-encoded, this is a nonissue in practice.)

Output the final sequence with print(1), since printf(1) is no longer
necessary for interpreting escape sequences or inserting the URI via
a conversion specification. I could have used `printf %s $seq`, but
idiomatic zsh prefers print(1).

  [1]: https://github.com/tmux/tmux/wiki/FAQ#what-is-the-passthrough-escape-sequence-and-how-do-i-use-it
  [2]: https://www.mail-archive.com/tmux-users@lists.sourceforge.net/msg03485.html