zsh: Clear terminal working directory if pwd fails
Fix a bug wherein if $PWD was null and pwd(1) failed, the local $PWD was
set to "x" instead of a null value.

Clear Terminal.app's current working directory if $PWD ends up null (now
that it can actually do so). Setting it to "file://" is harmless (the
proxy icon becomes a faded folder overlaid with a question mark), but
"file://hostname" and "file://user@hostname" are equivalent to the root
directory, which is not what a null $PWD indicates. Better to display no
directory at all.