gnupg, sh: Don't error-handle `export` or `shift`
Using `export` or `shift` on the left-hand side of an OR list is not
reliable because their errors may prevent the right-hand side from
running at all. POSIX.1-2017 XCU 2.14 [1] states that "[a]n error in
a special built-in utility may cause a shell executing that utility to
abort". Concerning interactive shells in particular, XCU 2.8.1 [2]
states that they shall not abort but instead "shall not perform any
further processing of the command in which the error occurred." Some
shells interpret this to include the remaining portions of an OR list,
including dash 0.5.11.3, ksh 93u+ 2012-08-01, and mksh R59c.

Avoid this for `export` by not error-checking it at all. If a variable
cannot be set or exported, there isn't much to be done anyway, other
than move on.

Avoid this for `shift` by checking for sufficient arguments beforehand.

  [1]: https://pubs.opengroup.org/onlinepubs/9699919799/utilities/V3_chap02.html#tag_18_14
  [2]: https://pubs.opengroup.org/onlinepubs/9699919799/utilities/V3_chap02.html#tag_18_08_01