promote(): Simplify handling of trivial calls
Invocations with no search arguments now print the list argument as-is
and return immediately. The previous behavior intended to produce the
same result by running the algorithm normally, but the implementation
ended up testing for the presence of search arguments anyway, so why not
just do that once at the beginning and be done with it.

Invocations with no arguments at all now return zero instead of nonzero.
This makes the code a little less fussy, and I don't really care about
changing the exit status because it's not obvious whether omitting all
arguments should be considered a failure condition in the first place.