terminfo: Stop setting TERM for Apple Terminal
When this topic script sets TERM to "nsterm-build400" and my
update_terminal_cwd zsh function is active, Terminal appears to restore
old values of the current working directory/document when certain
programs are run. This can be seen by starting a new terminal, running
and quitting less(1), using cd(1) to switch to another directory, and
running less(1) again. At this point Terminal's current working
directory is not the new directory but the one in which less(1) was run
initially. This also occurs with vim(1), nano(1), and no doubt other
software, but it does NOT occur when using less(1) via `git diff`.
I have no idea what is going on or how to even begin investigating.