macports: Stop using `sed(1)` to read path files
Replace the `sed | xargs2 promote` pipelines with a helper function that
reads a file, skips empty lines, and runs promote() in the caller's
shell environment.

This is preparation for modifying promote() to use an output variable so
it can be used without command substitution and the attendant subshells.
Doing that in a POSIX-compliant way while continuing to use sed(1) would
have been possible but ugly.  It wouldn't even have saved on subshells
because sed(1) itself would still have needed one.

This sort of reverts 3036132 (Move POSIX shell helper functions to
separate file, 2021-06-21).