macports: Only try to adjust nonempty PATH/MANPATH
In my current setup -- zsh 5.8.1 on macOS 10.14.6 (18G9323) -- MANPATH
seems to be always set, albeit empty; I assume the same is true of PATH.
I don't know whether this is due to zsh, my configuration, or macOS, but
it doesn't really matter because, in retrospect, testing whether these
variables were merely set was never correct.  After all, there'd be no
point in trying to adjust them if they were set but empty.

Switch to testing whether PATH and MANPATH are not empty.