sh: Rename functions library to _functions.sh
Using a distinct prefix makes it easier to skip the library while
sourcing the other scripts in ~/.profile.d.  Instead of manually
checking all the filenames, a simple glob now suffices.

(The source filename starts with two underscores because my existing
convention is that a lone leading underscore in a source pathname
component becomes a leading *period* in the installation.)