zsh: Move terminal color setup to a function
Moving this out of .zshrc gives me space to make this more complicated
than it has to be.

The new function speeds things up by using hard-coded control sequences
for terminals I use most often.  For other terminals, it first tries to
use zsh's loadable modules to query terminfo/termcap [*] and only falls
back to the external tput(1) utility as a last resort.

Doing all this in an autoloadable function makes it available outside of
startup, which could theoretically be useful.

(You might ask why I didn't just detect a few ANSI-compatible terminals,
use hard-coded sequences for those, and disable color otherwise, instead
of writing an overly flexible implementation with error-checking and the
works?  The answer is that that didn't occur to me until I was basically
done with this.)

[*]: Real talk: zsh/termcap support is currently entirely theoretical,
     as I haven't tested it.  I've left it commented out for now.