Unquote integers in `m4` templates
I began quoting integers in imitation of the GNU M4 manual, which does
so to maintain consistency [1]:

    For consistency, this manual follows the rule of thumb that each
    layer of parentheses introduces another layer of single quoting,
    except when showing the consequences of quoting rules.  This is done
    even when the quoted string cannot be a macro, such as with integers
    when you have not changed the syntax via `changeword` [...].

However, these days I care less about rote consistency and more about
clarity and concision.  Seeing unnecessary quoting makes me stop and
wonder why I quoted something that doesn't need to be quoted.  For me,
it reduces clarity.

Additionally, I don't have to worry about integers undergoing unwanted
macro expansions.  I try to conform to POSIX.1-2017 [2], which specifies
the composition of macro names -- letters, digits, and underscores, but
not starting with a digit -- and does not specify a way to change it.

  [1]: https://www.gnu.org/software/m4/manual/html_node/Quoting-Arguments.html
  [2]: https://pubs.opengroup.org/onlinepubs/9699919799/utilities/m4.html