Reformat shell conditionals
My preference these days is for each "important" command to be the first
visible content on its line, instead of obscured in the middle.  I think
this makes it easier to read and understand the code.  (I don't always
insist on this, but I'm not going to elaborate on the exceptions here.)