sh: Factor out most of promote() into partition()
The new partition() function resembles promote(), except that it puts
matching and nonmatching elements into separate variables.  This will
allow a future demote() to remain mostly DRY.

Now that the meat of promote() is in partition(), rewrite the former to
leverage the latter.  The new promote() should work the same as the old
one, except that when called without arguments it unsets REPLY instead
of giving it an empty value.

Expand the promote() comment to be more informative.