Unset `sh` variables with `unset -v`, not `unset`
POSIX.1-2017 does not specify the behavior of "unset foo" if a variable
named "foo" is not set [1].  In most modern shells [2] it does nothing
and has a zero exit status, but in bash [3] it tries to unset a function
named "foo".

I can't imagine this ever becoming a real problem here (several unlikely
conditions would have to be met), but using "unset -v foo" nips it in
the bud at practically no cost, so why not.

(There's no need to change the zsh-specific startup files because zsh
exhibits the common behavior, not the bash behavior.)

  [1]: https://pubs.opengroup.org/onlinepubs/9699919799/utilities/V3_chap02.html#unset
  [2]: I tested dash 0.5.9, dash 0.5.11.5, ksh93u+ 2012-08-01, mksh R59
       2020/10/31, mksh R59 2021/01/24, yash 2.52, zsh 5.3, and zsh 5.9.
  [3]: Since 2.05b, at least.