Don't let `unset_sh_helper_functions` unset itself
It was pretty convenient to have unset_sh_helper_functions unset itself
when it finished -- so convenient that it made ksh93 segfault:

    % ksh
    $ print -r -- "$KSH_VERSION"
    Version AJM 93u+ 2012-08-01
    $ . ~/.profile.d/_functions.sh
    $ typeset -f unset_sh_helper_functions
    unset_sh_helper_functions() {
        unset -f demote partition promote
        unset -f unset_sh_helper_functions
    }

    $ unset_sh_helper_functions
    zsh: segmentation fault  ksh
    %

I had some ideas for workarounds but decided to keep it simple and have
the callers do the unsetting.  There will always be a limited number of
callers, and I control all of them.