zsh: Source .profile as final command in .zprofile
When a fatal error (e.g., a variable assignment error) occurs in a zsh
startup file, evaluation of the file immediately terminates, and zsh
moves on to the next startup file.  However, a fatal error in a script
being sourced by a startup file only terminates the relevant dot/source
command; any commands that follow it are still executed.

This is not inherently bad.  The problem, insofar as there is one, is
that my .zprofile delegates more or less everything to .profile, which
it sources.  So an error that would abort any of the other startup files
does *not* abort .zprofile -- it only aborts the sourced proxy.  This
makes it annoying to think about errors, as my .zprofile basically plays
by its own set of rules.

Even things out by having .zprofile source .profile as its last command.
Now fatal errors in .profile *effectively* abort .zprofile, even if they
don't do so *technically*.