Export shell variables after verifying assignment
This prevents bash [1], yash [2], and recent zsh [3] from exporting
variables that they failed to properly set.  Only yash requires AND
lists here, but they accomplish the goal for bash and zsh as well.

Also: "Undo" 93378ae (lynx: Ensure LYNX_CFG contains home directory,
2022-03-15), which is unnecessary now that I'm setting LYNX_CFG with
a simple assignment.

(NB: It isn't actually necessary to fix the JAVA_HOME export because
that command can only be reached if JAVA_HOME is assignable in the first
place, but it's easier to be consistent than to justify a deviation.)

  [1]: Observed in 3.2.57 and 5.2.9.
  [2]: Observed in 2.52.
  [3]: Observed in 5.9 but *not* in 5.3.