Track temporary shell functions in a variable
The unset_sh_helper_functions function lost some of its shine when it
stopped self-deleting in f141979 (Don't let `unset_sh_helper_functions`
unset itself, 2023-02-26).  The new requirement that callers unset it
themselves was a new wart on top of existing warts:

  - Since it was defined at the end of _functions.sh, it was easy to
    forget to update it to include new functions.  (I don't think I ever
    did forget, but I could have, and that would have been annoying.)

  - It only unset other functions in _functions.sh, so the zsh startup
    files had to unset zsh-only functions separately.  (I'm sure I could
    have implemented something convoluted like a zsh-only function that
    invoked unset_sh_helper_functions and then unset the other zsh-only
    functions, but that sounds excessive.)

Considering all warts together, I decided to take another approach.

I now use a space-separated list to track the names of functions that
should be removed when the startup files wrap up.  As a variable, it can
be built up incrementally, allowing me to add each function right after
it is defined.  And the zsh startup files can use it too.

(One extra-new wart is that for portability I've implemented the list as
a scalar and thus rely on field splitting to use it.  Since I control
the variable contents there shouldn't be a problem, but it's still icky,
and I would have preferred an array.)