Return early on shell variable modification errors
This keeps lenient shells from plowing through functions and sourced
scripts after a variable assignment or unset(1) error has rendered the
subsequent code nonsensical.

(I don't currently feel like enumerating all the shells/situations this
applies to, but that would be more appropriate for a comment or a coding
guide in the repository.  For now I'll just note that I don't really
handle bash's behavior, which is in an annoying middle ground between
yash's "don't worry, be happy" leniency and dash's "everything is fatal"
strictness.  I might do so in the future, but it would require obtuse,
intrusive changes that I'm not sure are worth it.)